+++ src/authentication/authentication.ts
@@ -1,144 +1,13 @@
 import { z } from "zod";
 import * as Msal from "@azure/msal-browser";
 import { jwtDecode } from "jwt-decode";
-import { AxiosError } from "axios";
 import { useHistory } from "react-router-dom";
 import { useEffect } from "react";
 import { useQueryClient } from "@tanstack/react-query";
 import { OrganizationsDataApi } from "../api/generated";
 import { ADMIN_PANEL_RICHIESTE, DASHBOARD, LOGIN } from "../navigation/routes";
 import { SessionApi } from "../api/generated_public";
-import { makeStore } from "./authenticationStore";
-import {
-  localStorageInitialState,
-  localStorageSetup
-} from "./authenticationPersistance";
-import {
-  UserSession,
-  AdminSession,
-  CurrentSession
-} from "./authenticationState";
-
-export const authenticationStore = makeStore(localStorageInitialState);
-localStorageSetup(authenticationStore);
-
-const NONCE_DURATION = 15 * 60 * 1000;
-
-function setUserNonceByState(state: string, nonce: string) {
-  const data = authenticationStore.get();
-  authenticationStore.set({
-    ...data,
-    userNonceByState: {
-      ...data.userNonceByState,
-      [state]: { nonce, exp: Date.now() + NONCE_DURATION }
-    }
-  });
-}
-
-function getUserNonceByState(state: string): string | undefined {
-  const data = authenticationStore.get();
-  return data.userNonceByState[state]?.nonce;
-}
-
-function deleteUserNonceByState(state: string) {
-  const data = authenticationStore.get();
-  const { [state]: deleted, ...userNonceByState } = data.userNonceByState;
-  authenticationStore.set({ ...data, userNonceByState });
-}
-
-function setAdminNonceByState(state: string, nonce: string) {
-  const data = authenticationStore.get();
-  authenticationStore.set({
-    ...data,
-    adminNonceByState: {
-      ...data.adminNonceByState,
-      [state]: { nonce, exp: Date.now() + NONCE_DURATION }
-    }
-  });
-}
-
-function getAdminNonceByState(state: string): string | undefined {
-  const data = authenticationStore.get();
-  return data.adminNonceByState[state]?.nonce;
-}
-
-function deleteAdminNonceByState(state: string) {
-  const data = authenticationStore.get();
-  const { [state]: deleted, ...adminNonceByState } = data.adminNonceByState;
-  authenticationStore.set({ ...data, adminNonceByState });
-}
-
-function setUserSession(fiscal_code: string, session: UserSession) {
-  const data = authenticationStore.get();
-  authenticationStore.set({
-    ...data,
-    userSessionByFiscalCode: {
-      ...data.userSessionByFiscalCode,
-      [fiscal_code]: session
-    }
-  });
-}
-
-function setAdminSession(name: string, session: AdminSession) {
-  const data = authenticationStore.get();
-  authenticationStore.set({
-    ...data,
-    adminSessionByName: { ...data.adminSessionByName, [name]: session }
-  });
-}
-
-export function setCurrentSession(session: CurrentSession) {
-  const data = authenticationStore.get();
-  authenticationStore.set({
-    ...data,
-    currentSession: session
-  });
-}
-
-export function deleteSession(session: CurrentSession) {
-  const data = authenticationStore.get();
-  if (session?.type === "user") {
-    const { [session.userFiscalCode]: deleted, ...userSessionByFiscalCode } =
-      data.userSessionByFiscalCode;
-    authenticationStore.set({
-      ...data,
-      userSessionByFiscalCode
-    });
-  }
-  if (session?.type === "admin") {
-    const { [session.name]: deleted, ...adminSessionByName } =
-      data.adminSessionByName;
-    authenticationStore.set({
-      ...data,
-      adminSessionByName
-    });
-  }
-}
-
-export function getMerchantToken(): string {
-  const data = authenticationStore.get();
-  if (data.currentSession?.type === "user") {
-    const merchantFiscalCode = data.currentSession.merchantFiscalCode;
-    return (
-      data.userSessionByFiscalCode[
-        data.currentSession.userFiscalCode
-      ]?.merchants.find(
-        merchant => merchant.organization_fiscal_code === merchantFiscalCode
-      )?.token ?? ""
-    );
-  }
-  return "";
-  // returning empty, invalid or expired token here is fine since authentication errors are handled on401
-}
-
-export function getAdminToken(): string {
-  const data = authenticationStore.get();
-  if (data.currentSession?.type === "admin") {
-    return data.adminSessionByName[data.currentSession.name]?.token ?? "";
-  }
-  return "";
-  // returning empty, invalid or expired token here is fine since authentication errors are handled on401
-}
+import { authenticationStore } from "./authenticationStore";
 
 export function goToUserLoginPage() {
   const targetUri = process.env.ONE_IDENTITY_LOGIN_URI ?? "";
@@ -154,7 +23,7 @@ export function goToUserLoginPage() {
     nonce,
     redirect_uri
   }).toString()}`;
-  setUserNonceByState(state, nonce);
+  authenticationStore.setUserNonceByState(state, nonce);
   // using window.location.href intead of link hreh since nonce must be different on every click
   // eslint-disable-next-line functional/immutable-data
   window.location.href = targetUrl;
@@ -163,7 +32,7 @@ export function goToUserLoginPage() {
 export function goToAdminLoginPage() {
   const state = randomAlphaNumericString(16);
   const nonce = randomAlphaNumericString(16);
-  setAdminNonceByState(state, nonce);
+  authenticationStore.setAdminNonceByState(state, nonce);
   void AdminAccess.loginRedirect({
     scopes: ["openid", process.env.MSAL_CLIENT_ID as string],
     state,
@@ -229,17 +98,14 @@ async function onUserLoginRedirect() {
   const state = searchParams.get("state") ?? "";
   const code = searchParams.get("code") ?? "";
   try {
-    const nonce = getUserNonceByState(state);
+    const nonce = authenticationStore.getUserNonceByState(state);
     if (!nonce) {
       return { type: "not-executed" } as const;
     }
-    deleteUserNonceByState(state);
+    authenticationStore.deleteUserNonceByState(state);
     const userTokenResponse = await sessionApi.createJwtSessionToken({
       createJwtSessionTokenRequest: { requestType: "oi", nonce, code }
     });
-    if (userTokenResponse instanceof AxiosError) {
-      throw userTokenResponse;
-    }
     const { fiscal_code, first_name, last_name, exp } =
       userJWTPayloadSchema.parse(jwtDecode<unknown>(userTokenResponse.data));
     const merchantTokensReponse = await organizationsDataApi.getOrganizations({
@@ -247,10 +113,7 @@ async function onUserLoginRedirect() {
         Authorization: `Bearer ${userTokenResponse.data}`
       }
     });
-    if (merchantTokensReponse instanceof AxiosError) {
-      throw merchantTokensReponse;
-    }
-    setUserSession(fiscal_code, {
+    authenticationStore.setUserSession(fiscal_code, {
       token: userTokenResponse.data,
       first_name,
       last_name,
@@ -271,11 +134,11 @@ async function onAdminLoginRedirect() {
     if (!result) {
       return { type: "not-executed" } as const;
     }
-    const nonce = getAdminNonceByState(result.state ?? "");
+    const nonce = authenticationStore.getAdminNonceByState(result.state ?? "");
     if (!nonce) {
       throw new Error("Admin login nonce not found");
     }
-    deleteAdminNonceByState(result.state ?? "");
+    authenticationStore.deleteAdminNonceByState(result.state ?? "");
     const adminTokenResponse = await sessionApi.createJwtSessionToken({
       createJwtSessionTokenRequest: {
         requestType: "ad",
@@ -283,14 +146,11 @@ async function onAdminLoginRedirect() {
         nonce
       }
     });
-    if (adminTokenResponse instanceof AxiosError) {
-      throw adminTokenResponse;
-    }
     const { first_name, last_name, exp } = adminJWTPayloadSchema.parse(
       jwtDecode<unknown>(adminTokenResponse.data)
     );
     const name = `${first_name} ${last_name}`;
-    setAdminSession(name, {
+    authenticationStore.setAdminSession(name, {
       token: adminTokenResponse.data,
       first_name,
       last_name,
@@ -302,6 +162,14 @@ async function onAdminLoginRedirect() {
     console.error(error);
     return { type: "error" } as const;
   }
+}
+
+export function adminLogoutRedirect() {
+  void AdminAccess.logoutRedirect();
+}
+
+export function adminLogoutPopup() {
+  void AdminAccess.logoutPopup();
 }
 
 const userLoginRedirectPromise = onUserLoginRedirect();
@@ -319,7 +187,7 @@ export function useLoginRedirect() {
 
     void userLoginRedirectPromise.then(state => {
       if (state.type === "success") {
-        setCurrentSession({
+        authenticationStore.setCurrentSession({
           type: "user",
           userFiscalCode: state.fiscal_code
         });
@@ -330,7 +198,7 @@ export function useLoginRedirect() {
 
     void adminLoginRedirectPromise.then(state => {
       if (state.type === "success") {
-        setCurrentSession({
+        authenticationStore.setCurrentSession({
           type: "admin",
           name: state.name
         });
